LIBILS-176. Created "AbstractAlephOpResult" abstract class
Modified JAXB classes to extend abstract class.

Added "isError" method.

https://issues.umd.edu/browse/LIBILS-176
